fix: remove error pages and default backend
Update nginx-ingress helmrelease.yaml to prevent error masking

If an application was throwing any error that is in this list, it forwarded that error to the default backend. A lot of 405 requests where POST requests which gave an 500 error, this resulting in the default backend to handle those requests, but that only accepts GET request, thus transforming the errors in 405. So in reality the real errors where masked, and the applications couldn't handle the error properly resulting in more issues.